Change vector register to RegXMM type.
Vector registers are 128 bits not 64 bits. We are reuse
x64 register class so vector registers are implemented as
RegXMM however, xmm is a x64 type implemente as simple register
type. Maybe they can change this to Reg128 or implement register
class as a arch independent class.

For now we just reuse RegXMM.