Fix some EmitXForm call parameters order
As written on ISA, the e.g andi has the following order of parameters:
andi. ra, rs, ui

And the order of those in the binary instruction is:
rs, ra, ui

The rule is still "the first parameter is the destiny of the operator"
is true on the assembly form, not on the binary form.

After inverting that I also found out some instructions were actually
wrong! anyway, let's fix it