Fix andli and addli when immediate size is 16b < imm <= 32b
When running hhvm, an and assertion fails as the immediate 0x1fffffff is
being cropped with a imm.w() call (value is too big to fit).

Also took this chance to make the same kind of fix to addli.

Notes:
* andli is a logical operation, so __no__ extension of signal is
required (that's the reason why the 16bits values are or'd)

* addli is an arithmetical operation, so extension of signal is required
(that's the reason why the 16bits values are add'd)