Change load/store operands to MemoryRef.
The operands for load/store PPC64 instructions changed
to use MemoryRef. A MemoryRef is a data structure who
can represents many address modes.

This data structure is reused from X64 assembler. Note
that X64 has many different address modes, for that reason
we need to do some adjusts if we get a adress mode that we
can't represent in PPC64 like base+index*scale+disp.

In X64 we can have the following address modes:

    - Direct Operand: displacement
    - Indirect Operand: (base)
    - Base + Displacement: displacement(base)
    - (Index * Scale) + Displacement: displacement(,index,scale)
    - Base + Index + Displacement: displacement(base,index)
    - Base +(Index * Scale) + Displacement: displacement(base,
      index,scale)

While in PPC64 we have:

    - Direct Operand: displacement (Form-D)
    - Indirect Operand: (Base with displacement = 0) (Form-D)
    - Base + Index: Index(Base) (Form-X)