Create PatchMemoryOperands function.
We reuse data structure MemoryRef who supports many types of address
modes used by X64. In PPC64 we don't support scaled address modes so
if we have something like this we have to patch the memory operands
to perform the same operations.

PatchMemoryOperand functions gets a Vptr (which have a MemoryRef) and
and converts the scaled mode operands to a (index) register who can
used by PPC64 instructions. This function also converts the displacement
to a register in case we have a displacement > 16 bits.