Move r0 to a reserved register on abi
r0 is used on prologue/epilogue to store/restore the link register. On
ABI it should always be r0 so it'll be marked as reserved.

Furthermore this fix an issue on our register bank that r0 should not be
used when refering to its content for some instructions (e.g: addi,
which would use the r0 parameter as the number '0' rather than the
content of 'r0').