First implementation of push/pop vasm instructions
As an experimental approach (maybe this is simply too much for the
problem), I guarded all the code to be easily to remove afterwards by
the macro flag PPC64_HAS_PUSH_POP. Search that if the code is going to
be removed or definetely coded.

r31 stores the top of the stack address and it's initialized right
before the first push. It does this once for the chunk of vasm code to
be emited.

The algorithm is explained on (currently revision 9):
https://gist.github.com/gut/956d6431412aad0fc626

Also fix how call/ret use the temporary stack generated/restored by
pushMinCallStack and PopMinCallStack. Now only the call is responsible
for setting it and the ret should pick the LR from it to branch back.