Fix vasm call when there was no vasm push before
For the push/pop functionality, the call was changed to copy the stack
pointer from rstktop(), and not the rsp(), that means that rstktop()
needs to be initialized in order for that to work.

The rstktop() is only initialized if a push is performed.

This patch solves the issue by forcing the initialization prior to call