Added a baseless load to cgLdBindAddr.
This HHIR uses a loadqp vasm who uses RIP register. This
register is only avaliable on x64/x86. In PowerPC we don't
have this register or a equivalent (accessible) register to
get the next instruction pointer.

This instructions is used for optimize execution on x64 because
x64 does not support mov addr64, reg. But PowerPC can do this.

So instead make a complex thing to get the IP we just create a
preprocessor switch who instead call loadqp we call a baseless
load to get the 64 bit address.