Commonlize F and V registers as x86_64
As stated on ABI[1], page 37, the 32 Floating Point (FP) registers are
located in the first 32 VSRs and the 32 Vector registers are located in
the last 32 VSRs.

That's different than x86_64, which they overlap each other (XMMs type).
For the sake of using HHVM, which is based on x86_64 architecture, we
can also consider them to be the same group. Therefore the FP and Vector
types are interchangable on HHVM POV.

The reason behind this is the VRegDbl type, which is can implicitly
construct the RegXMM type. That means all of the FP operations need also
to use RegXMM type and not Reg64.

Another solution would be to create a new type so we can index these
registers, but that would need lots of changes in the common code.

[1] https://wiki.debian.org/ppc64el?action=AttachFile&do=view&target=ABI_v2_ppc64el.pdf