Implement stub_size for PPC64
Added stub_size return for PPC64. The stub size will be calculated
by (kNumArgs  + 2)* 4 + 8. So the length will be 32 bytes. X64 uses
(kNumargs + 2) * 10 + 7 where 10 is the maximum length of mov
instruction and 7 the maximum length for lea instruction. In PPC64
all instructions are 4 bytes so i guess the stub_size may have to be
at least 32 bytes, however it's just a guessing.