Fix tls usage on ppc64
DS means data segment. The variable seg is set as fs (a register that points to
thread memory content). Because fs != ds, the assert fails.

In x64 the function emitTLSAddr returns a Vptr address relative to FS register.

For ppc64, there is no fs register. To refer to a thread content, the ppc64-abi
uses a displacement relative to r13 (thread pointer).

Fix #81