Fix use of temp register, shift and or operations
When executing the function "VptrAddressToReg", the use of a register not
reserved could crash the code, since the register could have valid
information in the context.
This commit is a temporarily hack, which declares a new temp register in
the abi-ppc64. The correct way to do that is execute the a "lowering" for
each instruction that uses a temp register, where it can be requested by
"v.makeReg()".
Or operation "||" was replaced by "|", which makes the bit operation.
shlqi was removed, since "s.index" is invalid on that context.