Isolate performance asserts with a switch, for now
Checking if a hot code needs to be inside the same instruction cache
line should only be cared on performance analysis. Yes, I know we're
possibily emitting more code than we should, so we can't reach this
milestone now, but that is also no constraints at this development
phase.

With this marking it'll be easier to track down which code pieces are
not taken care.