Fix cmpbim. It can't be lowered to cmplim
If only 1 byte from memory should be compared, we can't read more than
that and use it on a bigger precision comparison. On ppc64, only 32bits
and 64bits data can be compared. The rest of the bits should be
sign-extended and the hardware is already doing that if a matching size
instruction is used.