Temporary fix for addqi
When emitting the addqi instruction calling the "emit" method, it does not
call the "lower" function for the VASM. If the immediate is greater than
16bits, the immediate must be loaded into a rewgister to perform the
operation.
This fix is temporary because the function "VptrToReg" need to be
refactored to do not use tmp registers, instead, it should use the dynamic
allocator as doen in the "lower" functions, i.e., "v.makeReg()".