Redefine start address to patch a bctr as 1st asm instruction
The previous code was setting the bctr instruction itself as the
beginning of the patch given address. This improves clarity as it's
pointing to the branch itself and can be easily validated, however the
ppc64_asm::Assembler is not the only one adding patch bctr.

The HPHP::jit::vasm_detail::emit_svcreq_stub is also adding addresses to
be patched and as we have no control on that (platform independent hhvm
code), we should adequate our patches to also point out the 1st
instruction a bctr branch would emit (remember, it emits a li64 before,
so it's a li/lis, and not a bctr)

ps: the patch bctr is not matching the li64 emitted instructions, this
will be fixed next by Igor.