Fix patchVptr when p.index is invalid
There is no point on vasm lowering to reuse registers, remember that
it's all virtual registers. Further optimizations will allocate
registers accordingly, and if an input is only used once, the
instruction that is using it might as well just output the same register
(the same may be used). e.g:

a = 1  // input
add:
b = a + 1 // output

If a is not used anywhere else, it can be optimized to

a = 1  // input
add:
a = a + 1 // output

but lowering is not the phase for that.

Now about the issue: as "tmp" was always using the p.index at the
beginning, it would use also if p.index is invalid, and that would be a
problem. Also calling tmp = v.makeReg() and then throwing that
assignment away was just a waste.