Enhancement: move patchVptr near to its usage
For debugging, when a storebi has no index and a displacement bigger
than 16bits, it makes more sense to see a:
ldimmb
ldimmq
storeb

being lowered than a (as currently):
ldimmq
ldimmb
storeb

That's because the ldimmq is related to the storeb addressing. If it
appears before it feels the storebi, which should only store a byte, is
actually processing quadword parameters.