Fix handleSRHelper's pointer to mcg structure
Currently it was setting the 1st arg as a pointer to the MCGenerator
structure:
ld r3, mcg@got(r2)

But what is actually intended is to set the 1st arg as the address of
the MCGenerator structure:
ld r3, mcg@got(r2)
ld r3, 0(r3)

It is like when you forget to dereference a pointer in C.

I guess @toc operator would eliminate the need to an additional ld, but
I got relocation overflow with that approach. Maybe a fix will come
later.