Fix ppc64's and x64's back-end.cpp to be architecture specific
It felt like the #ifdef for platform on back-end-x64.cpp should be used
for ppc64 detection but actually it is only used to handle different X64
ABIs. With JIT, ppc64 needs a different back-end file just like ARM
does. Only on interpreter mode that it was not needed.

This fix 02247501a599171f887a1ca8e3bb48728015a997 and consolidate the
back-end-x64.cpp to be as the master branch before our interpreter mode
got merged.