Refactor vasms' lowering (2/3)
Remove the big switch case by using VASM_OPCODES macro.

Provide lowering with only the needed: the given vasm that can be
lowered and a Vout instance that can be used to emit other vasms. If it
returns 'true', then the current vasm will be replaced by the new
emitted vasms.