Get PPC disassembly printing in printir:2
At this point it would be helpful to see what instructions
the assembler is emitting. This is a class will be used to
print instructions on trace. This will use decoder class.
Note this is different from Decoder class defined on asm-ppc
who is used by backend to identify a instruction or fields of
a instructions into a code block.