Update ppc64 files accordingly due to last merge
Update third-party but keep our folly branch (I updated it with the
latest content that hhvm's third-party is pointing to except for the
folly repository as it has non-merged pull requests)

Remove ppc64's back-end and updated all files that used to include it.

Update changes to relocation and unique-stubs as X64

Fix enterTCHelper that is setting TOC pointer incorrectly (Now this
function is not being called with the $r12 pointer carrying the function
address therefore the TOC pointer can't be restored. Researching for a
better solution...)

For some reason movb vasm is now const - It raises 'unimplemented vasm'
if the const modifier is not added

Update CALLEE_SAVED_BARRIER accordigly for ppc64