Push/pop now uses 8 bytes
Very simple push/pop, just like X64. The call, however, had to be
adapted to guarantee that no push/pop element is overwritten.

Also rearrange leavetc and call vasms as if a vasm reemits another vasm,
it should not be done in its definition or else the fallback template
function might be used.

Call vasms are now creating another stack in order to have the 32 bytes
minimal requirement of the ABI.