Fix tailcallphp vasm behavior
Remember the following trick on X64 for 'push + jmp':

It intends to manipulate a return address for the called function to
return. If it performed a simple 'call', then it would return to the
next instruction but a 'push + jmp' can make it be anywhere else.

With this trick, a tail call can avoid a routine like:

call
ret

and make the caller to return directly to the caller's return.

But for PPC64 I didn't find a way out, I'm making the basic one.