Adjust correctly the "move to" vasms that are effectful
Well, effectful is not a very nice word, but as documented on the
removeDeadCode header:
"instructions that mutate memory, physical registers, or status flags
are considered useful. All branches are considered useful."

Therefore a, e.g, mtlr is a dead-end for the Vreg but actually the LR
was changed and HHVM was removing that code as it didn't seem to change
anything.