Fix HHVM compilation issue due dasm-ppc64 on release mode.
Because of the new decoder HHVM was hanging compilation on release
mode in dasm-ppc64.cpp. This was occurring because gcc was trying to
do agresive optimizations on decoder-ppc64.cpp file. The DecoderTable
constructor have a huge initialization because he need to create a
table with all ppc64 instructions. But the constructor is a singleton
hence this is called only once so optimizations will not be a issue here.

So the optimization are dissabled for this construction only avoiding gcc
optimization here.

This commit also remove copy constructor and make class inherith from
boost::noncopyable. Copy constructors are unnecessary here.