Defining rone(r27) as "1"
VASMs of increment and decrement, must set the CR depending of the
operation result.
The instructions in PPC64 that we are using to emit these VASMs, like
"addi" (using immediate) does not affect the CR, so the other option
(implemented in this change) is use the "add" (using register), which has
the option to set the CR. For that, the register r27 was reserved and
initialized with value "1" to be used in these instructions.