Create code block for smashMovq
The method "mcg->code.blockFor(inst);" try to get the code block for the
instruction parameter, but this block is the same that is being written by
the generator, so it is locked.
As the same function for x64_64 writes in the instruction address, the
solution is create a code block pointing to the instruction and overwrite
it using the Assembler.