Get PPC disassembly printing in printir:2
This commit includes printing of ppc64 instruction operands. Still has
some details to work, like printing negative numbers, and branch targets.
This include a isNop function to verify if a instruction is a nop.
I suggest we all use this same way to verify the kind of a instruction
if is needed.

This commit also refactor some code. toString method now belongs to DecoderInfo.
Decode method now returns a DecoderInfo struct. This can be used to extract some
other informations for debug purpose. I also refactor the code to adapt it to
Facebook code standards.