Computing Status Flags (SF) everywhere where required
If a vasm has a VregSF member, it means it wants to modify the status
flags (our CR). As there is only one at the moment, we'll always use CR0
and don't need to check that.

There are some vasm though that needs to be reworked just like inc/dec.
For now they are just marked.