Adapt majority of immediates to Immed (32bits) only
There is only one kind of Immed that can be 64bits wide and that's the
one used by ldimmq. All other instructions that uise an immediate (like
addqi, cmpqi) all have an immediate of up to 32bits (due to X64
constraints).

This improves our generated code as li64 doesn't need to run that often.
We can change it to li32 and have fewer instructions and nops.

Also change patchImm for handling immediate as signed

ps: a workaround is implemented for the case that a negative immediate
is to be patched. I'm still looking for it...