Implement cmpsd for comparing VregDbl
And the vasms fcmpo, fcmpu, isel and mfcr.

This is unfortunately a workaround as ir-lower.h is highly directed to
X64. We should probably contact HHVM team and clarify if that can't be
shifted into X64's vasm lowering.

Here is how it works:
* cmpsd compares 2 floats and outputs on a 3rd float a result
  * cmpsd has an additional parameter to point out if it wants equal or
  unequal comparison results.
* The ir-lower.h transforms this result to a GP
* and then 'and' it with 1. Now this value can be used to analyse the
comparison.

Geez... re-doing that in PPC64 is just wrong! fcmpo and fcmpu already
return on CR bits the results!