Add const modifier to vasm emission
If a function parameter is defined by reference, then it's much faster
than to assign it to a new object as our vasms were being done.

(Though this is only possible when the parameter is not modified and on
X64 the prep/commute/binary operations were modifying the parameter.
That's the historical reason why we didn't have those const modifiers)

Inspired by: fdbd669f904f8588cece9d5806b83a6d930482d5