Solve and document FixupMap::fixupWork for ppc64
This solves exception handling for some cases (like division by zero).

Unfortunately the backchain problem can not be solved on PPC64 as there
is no way to influence the backchain address for a called function
other than having it in a contiguous address.

On X64 the $rbp is the one pointing the backchain, and it's simply
pushed to whatever position the $rsp is.

In the end, that's how the stack is being defined for division by zero:

PPC64:
------

0x3fffffffdd50: 0x3fffffffddd0
0x3fffffffdd60: 0x0000140e31ac          -> points to raiseWarning

0x3fffffffddd0: 0x3fffffffddf0 \ funcBodyHelper
0x3fffffffdde0: 0x000019400b38 / Frame  -> stashedAR->m_savedRip (fixup)

0x3fffffffddf0: 0x3fffab8bffb0
0x3fffffffde00: 0x000017190084          -> points to enterTCExit

0x3fffab8bffb0: 0x3fffffffdea0 \ VM
0x3fffab8bffc0: 0x0000194005a0 / Frame

0x3fffffffdea0: 0x3fffffffe030 \ enterTCHelper
0x3fffffffdeb0: 0x0000144ead84 /        -> points to enterTCImpl

X64:
----

0x7fffffffd330: 0x7fffffffd350
0x7fffffffd338: 0x00000429659d          -> points to raiseWarning

0x7fffffffd350: 0x7fffe187ffc0 \ funcBodyHelper
0x7fffffffd358: 0x000008400391 | Frame  -> stashedAR->m_savedRip (fixup)
0x7fffffffd360: 0x00000607cd19 |        -> points to enterTCExit
0x7fffffffd368: 0x7fffe2000008 /        -> rvmfp()

0x7fffe187ffc0: 0x7fffffffd370 \ VM
0x7fffe187ffc8: 0x0000084001d8 / Frame

0x7fffffffd370: 0x7fffffffd3d0 \ enterTCHelper
0x7fffffffd378: 0x000004f15dc4 /        -> points to enterTCImpl