Set SF registers when a vasm requires it
Generally it's not possible to use a ppc instruction that handles
immediate because they normally doesn't change the Condition Register,
which is required when a vasm has a VregSF parameter. This means we have
to change it to the register form so we can set CR.

By doing that, it could be noticed that the addq can't be used to
manipulate the rsp() (an assert in spEffect fail). The addqi still
exists for that purpose.

Also the lowerImm function was added to force immediate to reg
conversion. Only the patchVptr (which deals with memory addresses, so
not relevant to have CR registers) and the comparison vasms (because
they always set the CR) were left out with the original patchImm
handling.