Fix instructions that use part of registers
For PPC64 there are no instructions to perform operations with only part of
a register, differently of x86_64.
The VASMs that are composed of these kind of operations, must be lowered to
first extract the operator into a register and later perform operation using
the entire register, i. e., 64bits. If it needs a result, like subtract
operations, the destination register must be filled considering the size of
operation. For example:
Consider d register with the value "0x44582C24A50CAD2".
For a subl instruction (l means 32bits), the result must be placed in the
lower 32bits (X means the result value):
"0x44582C2XXXXXXXX" and the higher 32bits should not be touched.
For operands the treatment is analogous, only the lower 32bits are
considered to perform the operation.

Instructions fixed: movl, ldimmb, cmpb, cmpl, cmpbi, cmpli, subl, subbi,
subli, testb, testl, testbi, testli, xorb, xorl, xorbi, andb, andl and andbi.
VASMs created: movlk and clrlwi.