Force all VM calls to set the backchain to the VM frame
Several calls from the VM inspect the call stack in order to search the
VM frame. The VM frame and its saved LR should be matched regardless of
how many call frames the VM created so bringing this backchain to
latest frames will result in the same structure as X64.

Fix #93