Fix external calls backchaining to VM frames
Unfortunately the rsp() can't be trusted to be modified only by the call
vasm in HHVM. The rsp() is also modified by other operations which
breaks the assumption made on 261bb13f535e3da2bd1a45045d450346a7306181.
There is no reliable way of grabing the backchain of the rsp as the
code-gen manipulates it and the reference would be lost.

This solution reserves one of our numerous GP register to have always
the VM backchain value on it called rbackchain() as defined on
abi-ppc64.h. It's the same kind of approach with the rone() and it's,
for ppc64, a good idea to optimize constants that would need to be
calculated during runtime.

Fix #96.