Fix enterTCHelper to change return address if stashedAR is valid
If there is a valid stashedAR pointer calling enterTCHelper, then it
also carries the return address to the unique stub emitCallToExit. It
was not being used before. This idea came from the X64 approach.

Also improved emitCallToExit to check if the value is as expected.

Fix #99

ps: 364 out of the 1040 quick tests are passing now.