Fix analyze lea
Lea instruction that is a base for an instruction that has a Vptr as
argument, can be removed when the register is virtual, and scale is 1, and
the index is invalid. The code was considering that when the scale is 1,
the index is invalid, but for PPC64 the lower for lea sets scale to 1 and
index to a valid register.
Based on this, the if changed need to check if index is valid, otherwise
it will be ignored.