Add TOC save/restore on callExtern and made it unwindable
And as callExtern is changed, the smashable call parameters needed to be
updated and this chance was used to create a "epilogue size" parameter.

Quite often the call is referenced by a pointer, but as a call in ppc64
comprehends several instructions, it's often calculated by having a
pointer to the beginning or end. Either way the return address of the
call has to be calculated in order to match the LR value after call.

The newly created epilogue will assist in syncpoint and unwind vasms.