Solve 'Assertion `uint32_t(size()) <= MaxSize' failed.' error
On PPC64 the 'this' pointer (r3) was being overwritten by the unique
stub emitFreeLocalsHelpers. That happened because it was not pointed out
how many args were being used on that call. There is no issue on X64 as
no additional registers are needed on that context, which is not true on
PPC64.

Test used: hphp/test/quick/xml_entity_loader.php

Quick test results: 737 / 1040 tests passed (111 more since 60de747
change)