Fix frame unwinding to VM frame
After VM frame unwinding, the frame pointer was not correctly restored
as the external call created a new frame but it didn't have the chance
to restore it as the exception is being throw and it'll not have the
chance to return and restore it.

This issue didn't happen on other exceptions because the catcher was on
a C++ frame, so it would unwind it further and would not see this issue.

Landingpad is the clean up code for an exception, as described on:
https://mentorembedded.github.io/cxx-abi/abi-eh.html

Fix #104

Improved quick tests: 739 / 1040 tests pass (6 more since
dba828a6d48a068c0d9d54f780374d1d204ead29)

Test used: hphp/test/quick/sqrt.php