Save registers values in emitDecRefHelper
The unique stub function emitDecRefHelper is responsible for saving it's
registers used as arguments. As this code was copied from x64 code and the
instruction declm for PPC64 uses additional registers, it was losing the
values of these registers, when execution this instruction.
This issue was fixed adding a new scope and saving the registers with
"PhysRegSaver", as made in the same code some lines below.
The new scope is necessary since "PhysRegSaver" creates the pop instructions
the the destructor is called.