Fix ucomisd
The correct instruction for ucomisd is not dcmpu, but fcmpu.
The instruction dcmpu is not about "double comparison". It only
treats "Decimal floating point" comparison.

One other change was necessary to make this work: changing orders
of s0 and s1 in asm emmition. I don't know why this works. Requires
some investigation.

936/1040 tests passed.

One more test is passing, comparing to the last commit. But some
local tests that deals with big numbers and double comparison are
now passing.