Fix emitFunctionEnterHelper stub
This unique stub emits stublogue, which creates a frame that does not
follow the ActRet. It is a problem when calling the enterTCHelper from
a JITed code, since there is a frame validation that considers the PPC64
abi.
The fix was create a frame following the PPC64 abi when emiting the
stublogue and restore to the stub format before emit the stubret.
It seems not the best solution, maybe add all of it in the stublogue and
stubret and ask FB mantainers.
For now this will be our solution.

Passed quick tests: 1040 - 88 = 952