Fix srem vasm
The old implementation was not wrong at all. But hhvm has a
register allocation optimization that was not compatible with the
old code.

We used to have the following code:
void emit(const srem& i) {
a->divd(i.d, i.s0, i.s1);
a->mulld(i.d, i.d, i.s1);
a->subf(i.d, i.d, i.s0);

As it can be seen above, we used i.d as a result register for divd
and also as a operand for mulld. One could expect that i.s1 and i.d
would be allocated in different registers. But this is not what
happens in this case. They are allocated in the same register
causing a messy situation.

HHVM internals probably expects that vasms implementations don't
use destiny registers as operands in further instructions. Assuming
that, there is no problem in allocating the same register for i.s1
and i.d.

Lowering this vasm to use more register was enough fix it.

Passed quick tests: 1040 - 79 = 961