Fix test array_packed.php
This fix is a workaround, loadups needs to be lowered!

When emmiting the vasm, the operands should be patched, but they
are not. An issue will be created in order to fix this behavior
correctly.

It was also necessary to use lxvd2x in spite of lxvw4x.

For some reason,