Fix invalid frame pointer in fixupWork for ppc64
Tail call elimination needs to be disabled on fixup function as the
PPC64 frame pointer points to the bottom of the frame, hence it changes
depending of the locals stored on the stack and it might be in a
different position than fixupWork, which would mean that the frame
pointer grabbed inside of fixup would not point to a valid frame anymore
on fixupWork context.

As this behavior only happens when compiler optimization is enabled,
this issue doesn't happen on debug mode.

Fix #116
Passed quick tests on release mode: 1040 - 36 = 1004
(7 tests LESS than what is reported on 17a9088 but it was on debug mode)