Fix eq_int_str
As pointed in x64 isa, cvtsi2sd converts a signed integer double
word to a double floating point value. In order to make this work
it is neccessary to use fcfid in spite of fcfids, once this last
one converts a signed integer double to a single precision
floating point.