Fix treatment for NAN operands in ucomisd
As it can be seen in implCmpRelDbl (irlower-cmp.cpp), hhvm treats
unordered operands as if the result of the comparison (ucomisd)
returns a negative (or lower) result.

So, in order to ppc64 be compliant to this behavior, this commit
sets the "negative" bit whenever the SO (see ppc64-ISA page 159)
bit is set.