Fix float-overflow.php
The vasms  imul, subq, need to set the overflow bit. In order to
make this work, it was necessary to change the instructions they
implement.

The ppc64 architecture set the SO bit whenever an overflow (when
enabled) happens. Once SO is set, it needs to be unset explicitly.
Unsetting it, leads to emit two more instructions before a
"bsoctr". So some adaptation was necessary to be done in more parts
of the code.

Tests passed = 1040 - 27 = 1013