Adjust frame pointer of the frame
This unique stub creates a new frame pushing the info on the stack. The
frame pointer is one of the info saved, but it has an address higher than
the real top of the stack, so when pushing other values, the data on the
stack is overwritten.
To solve that, the frame pointer was saved before the frame being created
and restored on returning.

Tests passed = 1040 - 10 = 1030