Fix rvmpf from specific frame
The frame created in the unique-stub-ppc64.cpp:emitFunctionEnterHelper
sets the r31 with a not expected value, when loading the context in
unwind-itanium.cpp:tc_unwind_personality.
To fix that, when the registers $r1 and $r31 are equals, the correct value
for $r31 is in it's address.
This does not seems the best solution, and maybe the error is coming from
another code. This case must be better analyzed, but this is the solution
for now.

Tests failing:
hphp/test/quick/exception_many_locals.php
hphp/test/quick/fb_intercept.php
hphp/test/quick/intercept2.php
hphp/test/quick/profile/Setprofile_generic.php
hphp/test/quick/retval_backtrace.php

Fix #126