Fix flag's type of intercept from char to int8_t
Firstly, the char type is designed to hold an ASCII character and
arithmetic should be avoided.
Secondly, the sign of a char type is not defined on C standard. On X64,
the char expands to "signed char" however PPC64 expands it to "unsigned
char", which is a common source of mistake found when porting
applications.

Setting it as int8_t is a better option and solve all the issues pointed
out above.