Fix calling convention when GP after FP is needed
Register allocation on ppc64 is not easy: If a fixed argument function
uses FP arguments, the aligned GP arg needs also to be reserved.

Here, the number_format function ("double, int, int, int" arguments,
with storage buffer) should use for it's first int argument the $r5, not
the $r4 how it was as $r3 is the storage buffer and $r4 is reserved due
to $f1 being used for the double.

I tried to make it nicely before by removing simdArgs from other places,
but then there was a lea trying to move a GP to FP, odd... this one is
neater despite wasting (possibly) a copy to an unused GP register
argument.

Fix #123

Quick tests: still 2 tests failing:
  hphp/test/quick/fb_intercept.php
  hphp/test/quick/intercept2.php

All tests:    (13485/13485) (378 skipped, 107 failed)
  previously: (13456/13456) (378 skipped, 185 failed)