Disable empty catch optimization
After analyze why a landingpad instruction is necessary when returning
from the debugger, the empty catch block optimization was found and it is
the reason that on debug mode a landingpad is necessary when returning.
The optimization supposes that a landingpad instruction does not emit any
code, so it can be removed for optimization, for example.
As PPC64 is emiting code for this instruction, the optimization must not
occur for now, so a "if defined" was added to disable it.

100% of the quick tests are passing.