Fix unsigned comparison
This commit is a diff extracted from branch fix-unsigned-cmp.

Checkout fix-unsigned-cmp for more information.

Basically, when emitting a compare instruction (cmp) a second
instruction is emmited for logical comparison (cmpl). The result of
this one is saved on CR1 (the other is saved on CR0).

Also, when emmiting a instruction that sets CR0 (like and., add.,
etc), the result is copied to CR1.

When branching after a logical operation, CR1 is used in spite of
CR0.