Fix "unimplement instruction testqi" error.
Summary:

This commit solves a "unimplemented instruction testqi" on test
quick/debugger/break_async.php. The instruction was already
implemented but we lower this instruction to fix issues related
to de immediate size and reemit the instruction as testq.
A new commit:

https://github.com/facebook/hhvm/commit/4a905d4754290201ae2db64cfc6d4554ae6335b6

Simplify andqi+testq which causes instruction testqi to be reemited
after  lower and in POWER we never emit testqi instruction directly
hence we do not have a emit(testqi) which causes the error.

This fix creates a emit function for testqi instead lower the
instruction.

Passing 1024/1039 quick tests.