Add asserts to avoid losing data
Two kind of instructions for PPC64 discard some low bits in order to create
the assembly instruction. These bits are be useless and must be zero, but
if any one of them is set, it's losing information and and it means that
something is not right.
The assert verifies if this info is being lost.