Handle unaligned memory access
When creating a load assembly instruction, for PPC64, it is using ld,
which only allows 32bits aligned memory access.
In order to make an unaligned access, the instruction ldx must be used, as
it receives the index in a register, which was added in the emit, if the
displacement is no 32bits aligned.