Remove magic numbers
In the fixup.h file, the "+2" is the size of a x64 frame, which was
replaced by kNativeFrameSize, the fame size arch independent.
As now the fixup is referecing the correct frame size, the "+2" in the
unique-stub.cpp, that was added to complete the frame size for PPC64, was
removed. The "+1" is the frame offset for x64 to find the m_savedRip,
which was replaced by a generic function to get this offset.