Fixes build hang issue.
Summary:

After the commit bellow:

https://github.com/facebook/hhvm/commit/a96a9507c84586170120f7375c39a5b4cb85bf51

our build hanging on 58% (when try to compile decoder).
 __atributte optimize must be declared on constructor in .h file
otherwise it will be ignored by gcc. Without disabling optimizations
for decoder construction build will hang because we have a huge
initialization list on this constructor. Disable optimizations for
this constructor is not a big issue since this is a singleton class
and we do not take any advantage using a agressive optimization in
this case.