Create class for TOC management
In this approach in spite of manipulating TOC explicitly we create
a std_vector to implement TOC functionalities. This vector must be
created with fixed size once dynamic vector resizing/allocation
may change the vector's base address.

The vector also must persist in memory. So that it is necessary to
declare it as static.