Optimize emitted instructions based on upcoming branch
Add an analysis pass to the ppc optimisation. This pass tracks
anything that uses a ConditionCode (jcc, the compare and move
family, etc). It then checks to see if the ConditionCode uses an
_Unsigned branch or not.

It then annotates instructions that use "copyCR0toCR1", so that
the expensive copy sequence is only emitted if the branch is
going to test against CR1 (for unsigned comparisons). If the
upcoming comparison is signed, the copy is not emitted.

Also annotated are the compare instructions: they now only emit
the second logical compare operation if it is going to be used.

Together, these optimisations result in:
 - A ~9% improvement in MediaWiki performance
 - A ~26% improvement in the following microbenchmark:

for ($i=0; $i < $count; $i++) {
	if ($i == -1) {
	} elseif ($i == -2) {
	} else if ($i == -3) {
	}
}

This patch has lots of debugging code still in it. The next
patch cleans that up. However, the whole approach of annotating
instruction structs might need to be revisited: it's quite
intrusive and upstream might prefer something that has less impact
on the generic code.

Signed-off-by: Daniel Axtens <dja@axtens.net>