More sophisticated handling of branch signedness
Instead of a bool needUnsigned, create an enum for the 4 different
states we could have:
 - signedOnly
 - unsignedOnly
 - neither
 - both
Use this to only emit the exact cmp/copy required.

Not a significant speed boost - some things slightly faster, others
slightly slower, all within the margin of error. I'm still trying
to figure out why that is. Assuming this could be made to work, it
could be used to clean up some more of the copying of CR0. As it
currently stands there is now very few of those copies emitted for
MediaWiki, so I'm leaving it for now.

Signed-off-by: Daniel Axtens <dja@axtens.net>