Fix r1 recovery on stubret
When leaving from tc on TcExit and entering the cpp code, r1 must
have the same value when it first entered tc on stublogue/inittc.
This value is saved on r31.

This commit fixes the problem pointed above, but the value saved on
r1 (on TcExit) points to itself, when it needs to point to the
previous frame.

Probably, this commit is not the best solution for the issue.
@lbianc , please review it.