Improvement on decoder search instruction
As the instruction list is placed statically on the code, it was sorted
based on its opcode to make the search easier on runtime.
The instruction opcode (6 first bits) is the only information placed on the
same place for all instructions, so based on that, an index was created to
help when search the instruction info.
As the list is ordered and grouped by opcode (6 first bits), the other
bits are being used in a binary search to find the instruction when
decoding.

Performance improvement: from 2,5% to 5% on OSS fibonacci.