RelocationSize should be the same for all platforms
It can't be either too big nor too small otherwise it'll simply not
relocate. Let us stick with 1mb and fix issues we find.

Only if the following error raises then it'll be allowed to set it to a
bigger size:

message: Attempted to emit 4 byte(s) into a 64 byte DataBlock with 0
bytes available. This almost certainly means the TC is full. If this is
the case, increasing Eval.JitASize, Eval.JitAColdSize,
Eval.JitAFrozenSize and Eval.JitGlobalDataSize in the configuration file
when running this script or application should fix this problem.