Remove unused BranchType on Assembler's patch
As our decoder is now robust, it's not needed to carry on this
information any further. It could be used to double check but in the end
the only difference between the BranchType is that a FarBranch has 2
more bytes when it's conditional, which would only mean more bytes to
read.